Refactor home page sections
