+++ components/pages/home/StatsSection.tsx
@@ -1,0 +1,12 @@
+export default function StatsSection() {
+  return (
+    <section className="py-16 px-6 text-center">
+      <h2 className="text-2xl font-bold mb-4">Une vue d'ensemble du territoire</h2>
+      <p className="text-gray-600 mb-8">+100 000 hectares analysés • +50 cultures référencées • Données publiques consolidées</p>
+      <div className="w-full max-w-4xl h-64 mx-auto bg-gray-200 rounded-lg shadow-inner flex items-center justify-center">
+        <span className="text-gray-500">[Mini carte interactive ici]</span>
+      </div>
+    </section>
+  );
+}
+
